decoupled_wbc/pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel", "pip"]
build-backend = "setuptools.build_meta"

[project]
name = "decoupled_wbc"
dynamic = ["version"]
readme = "../README.md"
classifiers = [
    "Intended Audience :: Science/Research",
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
authors = [
    {name = "NVIDIA Gear Lab"}
]
requires-python = "~=3.10.0"
dependencies = [
    "numpy==1.26.4",
    "scipy==1.15.3",
    "torch",
]
license = {file = "../LICENSE"}

[project.optional-dependencies]
# Full: all dependencies for the complete decoupled_wbc project
# Usage: pip install -e "decoupled_wbc[full]"
full = [
    "av>=14.2",
    "pyttsx3==2.90",
    "matplotlib",
    "hydra-core",
    "ray[default]",
    "click",
    "gymnasium",
    "mujoco",
    "termcolor",
    "flask",
    "python-socketio>=5.13.0",
    "flask_socketio",
    "loguru",
    "meshcat",
    "meshcat-shapes",
    "onnxruntime",
    "rerun-sdk==0.21.0",
    "pygame",
    "sshkeyboard",
    "msgpack",
    "msgpack-numpy",
    "pyzmq",
    "PyQt6; platform_machine != 'aarch64'",
    "pin",
    "pin-pink",
    "pyrealsense2; sys_platform != 'darwin'",
    "pyrealsense2-macosx; sys_platform == 'darwin'",
    "qpsolvers[osqp,quadprog]",
    "tyro",
    "cv-bridge",
    "lark",
    "lerobot @ git+https://github.com/huggingface/lerobot.git@a445d9c9da6bea99a8972daa4fe1fdd053d711d2",
    "datasets==3.6.0",
    "pandas",
    "evdev; sys_platform == 'linux'",
    "pyyaml",
]
dev = [
    "pytest==7.4.0",
    "build",
    "setuptools",
    "wheel",
    "ruff",
    "black",
    "ipdb",
]

[project.scripts]
decoupled_wbc = "decoupled_wbc.control.teleop.gui.cli:cli"

[tool.setuptools.packages.find]
where = [".."]
include = ["decoupled_wbc*"]

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
decoupled_wbc = ["py.typed", "**/*.json", "**/*.yaml"]

[tool.setuptools.dynamic]
version = {attr = "decoupled_wbc.version.VERSION"}

